Move HTTP parser into its own crate
